streamlit/requirements.txt
altair==4.2.0 ; python_full_version == "3.9.15"
attrs==22.1.0 ; python_full_version == "3.9.15"
blinker==1.5 ; python_full_version == "3.9.15"
cachetools==5.2.0 ; python_full_version == "3.9.15"
certifi==2022.9.24 ; python_full_version == "3.9.15"
charset-normalizer==2.1.1 ; python_full_version == "3.9.15"
click==8.1.3 ; python_full_version == "3.9.15"
colorama==0.4.6 ; platform_system == "Windows" and python_full_version == "3.9.15"
commonmark==0.9.1 ; python_full_version == "3.9.15"
confluent-kafka==1.9.2 ; python_full_version == "3.9.15"
decorator==5.1.1 ; python_full_version == "3.9.15"
entrypoints==0.4 ; python_full_version == "3.9.15"
gitdb==4.0.10 ; python_full_version == "3.9.15"
gitpython==3.1.29 ; python_full_version == "3.9.15"
idna==3.4 ; python_full_version == "3.9.15"
importlib-metadata==5.1.0 ; python_full_version == "3.9.15"
jinja2==3.1.2 ; python_full_version == "3.9.15"
jsonschema==4.17.1 ; python_full_version == "3.9.15"
markupsafe==2.1.1 ; python_full_version == "3.9.15"
numpy==1.23.5 ; python_full_version == "3.9.15"
packaging==21.3 ; python_full_version == "3.9.15"
pandas==1.5.2 ; python_full_version == "3.9.15"
pillow==9.3.0 ; python_full_version == "3.9.15"
protobuf==3.20.3 ; python_full_version == "3.9.15"
pyarrow==10.0.1 ; python_full_version == "3.9.15"
pydeck==0.8.0 ; python_full_version == "3.9.15"
pygments==2.13.0 ; python_full_version == "3.9.15"
pympler==1.0.1 ; python_full_version == "3.9.15"
pyparsing==3.0.9 ; python_full_version == "3.9.15"
pyrsistent==0.19.2 ; python_full_version == "3.9.15"
python-dateutil==2.8.2 ; python_full_version == "3.9.15"
python-dotenv==0.21.0 ; python_full_version == "3.9.15"
pytz-deprecation-shim==0.1.0.post0 ; python_full_version == "3.9.15"
pytz==2022.6 ; python_full_version == "3.9.15"
requests==2.28.1 ; python_full_version == "3.9.15"
rich==12.6.0 ; python_full_version == "3.9.15"
semver==2.13.0 ; python_full_version == "3.9.15"
six==1.16.0 ; python_full_version == "3.9.15"
smmap==5.0.0 ; python_full_version == "3.9.15"
streamlit==1.15.1 ; python_full_version == "3.9.15"
toml==0.10.2 ; python_full_version == "3.9.15"
toolz==0.12.0 ; python_full_version == "3.9.15"
tornado==6.2 ; python_full_version == "3.9.15"
typing-extensions==4.4.0 ; python_full_version == "3.9.15"
tzdata==2022.6 ; python_full_version == "3.9.15"
tzlocal==4.2 ; python_full_version == "3.9.15"
urllib3==1.26.13 ; python_full_version == "3.9.15"
validators==0.20.0 ; python_full_version == "3.9.15"
watchdog==2.1.9 ; python_full_version == "3.9.15"
zipp==3.11.0 ; python_full_version == "3.9.15"
